datetime picker, need styling

+++ src/bccf/models.py
@@ -1014,8 +1014,9 @@ class Event(BCCFChildPage):
                 variation.unit_price = self.price
                 variation.save()
         else:
-            self.event_product.delete()
-            self.event_product = None
+            if self.event_product:
+                self.event_product.delete()
+                self.event_product = None
                 
         super(Event, self).save(**kwargs)
